Conflicts:
	Juega-V2/Juega.v12.suo
	Juega-V2/Juega/Controllers/Juega/PerfilController.cs
	Juega-V2/Juega/Models/Juega/UsuarioModel.cs
	Juega-V2/Juega/Scripts/_references.js
	Juega-V2/Juega/Web.config
	Juega-V2/Juega/bin/Juega.dll
	Juega-V2/Juega/bin/Juega.dll.config
	Juega-V2/Juega/bin/Juega.pdb
	Juega-V2/Juega/obj/Debug/DesignTimeResolveAssemblyReferencesInput.cache
	Juega-V2/Juega/obj/Debug/Juega.csprojResolveAssemblyReference.cache
	Juega-V2/Juega/obj/Debug/Juega.dll
	Juega-V2/Juega/obj/Debug/Juega.pdb

+++ Juega-V2/Juega/BDD/Cancha.cs
@@ -27,7 +27,7 @@ namespace Juega.BDD
         public Nullable<int> NumEspectadores { get; set; }
         public Nullable<int> Largo { get; set; }
         public Nullable<int> Ancho { get; set; }
-        public Nullable<int> Valoracion { get; set; }
+        public Nullable<decimal> Valoracion { get; set; }
         public string TipoEstado { get; set; }
         public string FotoPrincipal { get; set; }
         public Nullable<System.DateTime> FechaCreo { get; set; }
